+++ lib/src/controllers/senzu_player_bundle.dart
@@ -106,9 +106,6 @@ class SenzuPlayerBundle {
 
   void dispose() {
     log('SENZU PLAYER BUNDLE DESPOSE');
-    // Core-с хамаарах controllerуудыг эхлээд хаана,
-    // дараа нь core-г хаана. Ингэснээр тэдгээр нь
-    // аль хэдийн устгагдсан core руу хандахгүй.
     sleepTimer.onClose();
     annotation.onClose();
     device.onClose();
@@ -117,6 +114,6 @@ class SenzuPlayerBundle {
     ad.onClose();
     subtitle.onClose();
     playback.onClose();
-    core.onClose(); // хамгийн сүүлд
+    core.onClose();
   }
 }